requirements.txt
beautifulsoup4==4.12.2
black==21.12b0
bs4==0.0.1
chromedriver-autoinstaller==0.4.0
chromedriver-binary==114.0.5735.90.0
flake8==6.0.0
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.3
pandas==2.0.2
pytest==7.3.1
requests-oauthlib==1.3.1
selenium==4.7.2
simplejson==3.19.1
webdriver-manager==3.8.6

